Begin joinClass and joinClassroom HTTP apis
